refactor: more code splitting
