+++ packages/application-shell/src/components/application-shell-authenticated/index.tsx
@@ -1,0 +1,10 @@
+import { lazy } from 'react';
+
+const ApplicationShellAuthenticated = lazy(
+  () =>
+    import(
+      './application-shell-authenticated' /* webpackChunkName: "application-shell-authenticated" */
+    )
+);
+
+export default ApplicationShellAuthenticated;
