+++ packages/application-shell/src/components/application-shell-provider/use-coerced-environment-values.ts
@@ -1,6 +1,5 @@
-import type { TApplicationContext } from '@commercetools-frontend/application-shell-connectors';
-
 import { useMemo } from 'react';
+import type { TApplicationContext } from '@commercetools-frontend/application-shell-connectors';
 
 type ShallowJson = { [key: string]: unknown };
 
@@ -45,16 +44,14 @@ const shallowlyCoerceValues = (uncoercedEnvironmentValues: ShallowJson) =>
     {}
   );
 
-const useCoercedEnvironmentValues = <
-  AdditionalEnvironmentProperties extends {}
->(
+const useCoercedEnvironmentValues = (
   environment: ShallowJson
-): TApplicationContext<AdditionalEnvironmentProperties>['environment'] => {
+): TApplicationContext<{}>['environment'] => {
   const coercedEnvironmentValues = useMemo(
     () =>
       shallowlyCoerceValues(
         environment
-      ) as TApplicationContext<AdditionalEnvironmentProperties>['environment'],
+      ) as TApplicationContext<{}>['environment'],
     [environment]
   );
   return coercedEnvironmentValues;
