+++ packages/application-shell/src/components/authenticated/oidc-callback.tsx
@@ -11,7 +11,7 @@ import OidcCallbackErrorPage from './oidc-callback-error-page';
 
 declare let window: ApplicationWindow;
 
-export type TProps = {
+export type TOidcCallbackProps = {
   locale: string;
   applicationMessages: TAsyncLocaleDataProps['applicationMessages'];
   children?: never;
@@ -19,7 +19,7 @@ export type TProps = {
 type AuthorizeCallbackFragments = { sessionToken?: string; state: string };
 type SessionToken = { nonce: string };
 
-const OidcCallback = (props: TProps) => {
+const OidcCallback = (props: TOidcCallbackProps) => {
   const location = useLocation();
   let errorMessage: string | undefined;
