+++ packages/application-shell/src/components/error-apologizer/index.ts
@@ -1,1 +1,7 @@
-export { default } from './error-apologizer';
+import { lazy } from 'react';
+
+const ErrorApologizer = lazy(
+  () => import('./error-apologizer' /* webpackChunkName: "error-apologizer" */)
+);
+
+export default ErrorApologizer;
