+++ packages/application-shell/src/components/fetch-project/fetch-project.tsx
@@ -15,13 +15,13 @@ type RenderFnArgs = {
   error?: ApolloError;
   project: TFetchProjectQuery['project'];
 };
-type Props = {
-  projectKey: string;
+type TFetchProjectProps = {
+  projectKey?: string;
   skip?: boolean;
   children: (args: RenderFnArgs) => ReactNode;
 };
 
-const FetchProject = (props: Props) => {
+const FetchProject = (props: TFetchProjectProps) => {
   const { loading, data, error } = useMcQuery<
     TFetchProjectQuery,
     TFetchProjectQueryVariables
