+++ packages/application-shell/src/components/project-container/index.ts
@@ -1,1 +1,8 @@
-export { default } from './project-container';
+import { lazy } from 'react';
+
+const ProjectContainer = lazy(
+  () =>
+    import('./project-container' /* webpackChunkName: "project-container" */)
+);
+
+export default ProjectContainer;
