+++ packages/application-shell/src/components/quick-access/quick-access.tsx
@@ -117,7 +117,9 @@ const QuickAccess = (props: Props) => {
     customApplications: true,
     canViewDashboard: true,
   });
-  const applicationContext = useApplicationContext();
+  const applicationContext = useApplicationContext<{
+    useFullRedirectsForLinks?: boolean;
+  }>();
 
   // Destructure functions from props to reference them in the hook dependency list
   const { onPimIndexerStateChange: onPimIndexerStateChangeFromParent } = props;
