+++ packages/application-shell/src/components/redirect-to-logout/redirect-to-logout.tsx
@@ -1,12 +1,13 @@
-import type { ApplicationWindow } from '@commercetools-frontend/constants';
-
 import { useLocation } from 'react-router-dom';
-import { LOGOUT_REASONS } from '@commercetools-frontend/constants';
+import {
+  LOGOUT_REASONS,
+  type ApplicationWindow,
+} from '@commercetools-frontend/constants';
 import useIsServedByProxy from '../../hooks/use-is-served-by-proxy';
 import * as oidcStorage from '../../utils/oidc-storage';
+import getMcOrigin from '../../utils/get-mc-origin';
 import Redirector from '../redirector';
-import RedirectToLogin from './redirect-to-login';
-import { getMcOrigin } from './helpers';
+import RedirectToLogin from '../redirect-to-login';
 
 declare let window: ApplicationWindow;
