+++ packages/application-shell/src/components/redirect-to-project-create/index.ts
@@ -1,1 +1,10 @@
-export { default } from './redirect-to-project-create';
+import { lazy } from 'react';
+
+const RedirectToProjectCreate = lazy(
+  () =>
+    import(
+      './redirect-to-project-create' /* webpackChunkName: "redirect-to-project-create" */
+    )
+);
+
+export default RedirectToProjectCreate;
