+++ packages/application-shell/src/components/requests-in-flight-loader/index.ts
@@ -1,2 +1,11 @@
-export { default } from './requests-in-flight-loader';
+import { lazy } from 'react';
+
+const RequestsInFlightLoader = lazy(
+  () =>
+    import(
+      './requests-in-flight-loader' /* webpackChunkName: "requests-in-flight-loader" */
+    )
+);
+
+export default RequestsInFlightLoader;
 export { default as requestsInFlightReducer } from './reducer';
