+++ packages/application-shell/src/components/route-catch-all/index.ts
@@ -1,1 +1,7 @@
-export { default } from './route-catch-all';
+import { lazy } from 'react';
+
+const RouteCatchAll = lazy(
+  () => import('./route-catch-all' /* webpackChunkName: "route-catch-all" */)
+);
+
+export default RouteCatchAll;
