+++ packages/application-shell/src/components/user-settings-menu/index.ts
@@ -1,1 +1,8 @@
-export { default } from './user-settings-menu';
+import { lazy } from 'react';
+
+const UserSettingsMenu = lazy(
+  () =>
+    import('./user-settings-menu' /* webpackChunkName: "user-settings-menu" */)
+);
+
+export default UserSettingsMenu;
