+++ packages/application-shell/src/components/version-tracker/index.ts
@@ -1,1 +1,7 @@
-export { default } from './version-tracker';
+import { lazy } from 'react';
+
+const VersionTracker = lazy(
+  () => import('./version-tracker' /* webpackChunkName: "version-tracker" */)
+);
+
+export default VersionTracker;
